requirements.txt
audioread==2.1.6
bleach==1.5.0
certifi==2018.8.24
cffi==1.12.1
cycler==0.10.0
decorator==4.3.2
enum34==1.1.6
future==0.17.1
h5py==2.9.0
html5lib==0.9999999
joblib==0.13.1
jsonschema==2.6.0
Keras==2.0.8
kiwisolver==1.0.1
librosa==0.6.2
llvmlite==0.27.0
Markdown==3.0.1
matplotlib==3.0.2
musdb==0.2.3
museval==0.2.0
numba==0.42.0
numpy==1.16.1
pescador==2.0.1
protobuf==3.6.1
pyaml==18.11.0
pycparser==2.19
pydub==0.23.1
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
pyzmq==17.1.2
resampy==0.2.1
scikit-learn==0.20.2
scipy==1.2.1
simplejson==3.16.0
six==1.12.0
SoundFile==0.10.2
stempeg==0.1.4
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0
tqdm==4.31.1
Werkzeug==0.14.1
